[DNSRSLVR] Use GetTickCount() in negative cache to fix older-SDK builds
The negative DNS cache uses GetTickCount64 for expiry timestamps but
that symbol isn't always present at link time on the older Windows
SDKs (MSVC ARM 14, RosBE gcc 8.4 targeting Vista 0x600 etc.) used by
ReactOS CI, leading to "undefined reference to GetTickCount64" link
failures across multiple build matrix configurations.

Switch to 32-bit GetTickCount(): the negative cache TTL is 15 minutes,
well below the ~24-day wraparound point, so a wraparound-safe
unsigned-subtraction check ((LONG)(NowMs - ExpireTickMs) >= 0) gives
correct expiry behavior without needing the Vista+ 64-bit tick API.

+++ base/services/dnsrslvr/precomp.h
@@ -52,7 +52,11 @@ typedef struct _RESOLVER_NEG_CACHE_ENTRY
     LPWSTR Name;
     WORD wType;
     DNS_STATUS Status;
-    ULONGLONG ExpireTickMs;
+    /* GetTickCount() value at which the entry expires.  Stored as DWORD
+     * to avoid pulling in GetTickCount64 (Vista+ only and not present in
+     * older Windows SDKs); time-since checks use the wraparound-safe
+     * unsigned subtraction pattern. */
+    DWORD ExpireTickMs;
 } RESOLVER_NEG_CACHE_ENTRY, *PRESOLVER_NEG_CACHE_ENTRY;
 
 typedef struct _RESOLVER_NEG_CACHE
